requirements_dev.txt
annotated-types==0.7.0
kafka-python==2.0.2
mypy==1.10.1
mypy-extensions==1.0.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
ruff==0.5.1
typing_extensions==4.12.2
